refactoring student profile data
