+++ server/Controllers/getStudentProfileData.js
@@ -1,7 +1,7 @@
 //I want my API to accept a student ID 
 //and give me total messages in each channel 
 
-import { studentProfileData } from "../slackMethods";
+import { studentProfileData } from "./studentProfileData";
 
 const getStudentProfileData = async (req,res) =>{
     const studentID = req.params.id;
